refactor out TypeBuilderUtils
