+++ packages/quicktype-core/src/attributes/StringTypes.ts
@@ -13,8 +13,7 @@ import {
 import { type DateTimeRecognizer } from "../DateTime";
 import { assert, defined } from "../support/Support";
 import { type TransformedStringTypeKind } from "../Type/TransformedStringType";
-// eslint-disable-next-line import/no-cycle
-import { type StringTypeMapping, stringTypeMappingGet } from "../TypeBuilder";
+import { type StringTypeMapping, stringTypeMappingGet } from "../Type/TypeBuilderUtils";
 
 import { TypeAttributeKind } from "./TypeAttributes";
