+++ packages/quicktype-core/src/index.ts
@@ -74,7 +74,8 @@ export { readableFromFileOrURL, readFromFileOrURL } from "./input/io/NodeIO";
 export { FetchingJSONSchemaStore } from "./input/FetchingJSONSchemaStore";
 export { JSONSchemaStore, type JSONSchema } from "./input/JSONSchemaStore";
 export { sourcesFromPostmanCollection } from "./input/PostmanCollection";
-export { TypeBuilder, type StringTypeMapping } from "./Type/TypeBuilder";
+export { TypeBuilder } from "./Type/TypeBuilder";
+export { type StringTypeMapping } from "./Type/TypeBuilderUtils";
 export { type TypeRef, derefTypeRef } from "./Type/TypeGraph";
 export { TypeAttributeKind, type TypeAttributes, emptyTypeAttributes } from "./attributes/TypeAttributes";
 export { TypeNames, makeNamesTypeAttributes, namesTypeAttributeKind } from "./attributes/TypeNames";
