+++ packages/quicktype-core/src/language/Swift/language.ts
@@ -4,7 +4,7 @@ import { BooleanOption, EnumOption, StringOption, getOptionValues } from "../../
 import { AcronymStyleOptions, acronymOption } from "../../support/Acronyms";
 import { TargetLanguage } from "../../TargetLanguage";
 import { type PrimitiveStringTypeKind, type TransformedStringTypeKind } from "../../Type";
-import { type StringTypeMapping } from "../../TypeBuilder";
+import { type StringTypeMapping } from "../../Type/TypeBuilderUtils";
 import { type FixMeOptionsType } from "../../types";
 
 import { SwiftRenderer } from "./SwiftRenderer";
